#12 updated cloud functions
